# Parallelize batched blob kzg proof verification