--- a/Arkworks/Cargo.toml
+++ b/Arkworks/Cargo.toml
@@ -16,6 +16,7 @@ merkle_light = "0.4.0"
 rand = "0.8"
 ark-std = { version = "^0.3.0", default-features = false }
 rayon = { version = "1", optional = true }
+num_cpus = "1.15.0"
 
 [dev-dependencies]
 criterion = "0.4.0"