--- a/Arkworks/src/eip_4844.rs
+++ b/Arkworks/src/eip_4844.rs
@@ -16,6 +16,9 @@ use kzg::{Poly, G1};
 use std::fs::File;
 use std::io::Read;
 
+#[cfg(feature = "parallel")]
+use rayon::prelude::*;
+
 pub fn bytes_to_bls_field(bytes: &[u8; BYTES_PER_FIELD_ELEMENT]) -> Result<FsFr, u8> {
     FsFr::from_bytes(*bytes)
 }
@@ -354,24 +357,37 @@ pub fn verify_blob_kzg_proof_batch(
         return verify_blob_kzg_proof(&blobs[0], &commitments_g1[0], &proofs_g1[0], ks);
     }
 
-    let mut evaluation_challenges_fr: Vec<FsFr> = Vec::new();
-    let mut ys_fr: Vec<FsFr> = Vec::new();
+    #[cfg(feature = "parallel")]
+    {
+        let results: Vec<bool> = (blobs, commitments_g1, proofs_g1)
+            .into_par_iter()
+            .map(|(blob, commitment, proof)| verify_blob_kzg_proof(blob, commitment, proof, ks))
+            .collect();
 
-    for i in 0..blobs.len() {
-        let polynomial = blob_to_polynomial(&blobs[i]);
-        let evaluation_challenge_fr = compute_challenge(&blobs[i], &commitments_g1[i]);
-        let y_fr =
-            evaluate_polynomial_in_evaluation_form(&polynomial, &evaluation_challenge_fr, ks);
-
-        evaluation_challenges_fr.push(evaluation_challenge_fr);
-        ys_fr.push(y_fr);
+        results.into_iter().all(|x| x)
     }
 
-    verify_kzg_proof_batch(
-        commitments_g1,
-        &evaluation_challenges_fr,
-        &ys_fr,
-        proofs_g1,
-        ks,
-    )
+    #[cfg(not(feature = "parallel"))]
+    {
+        let mut evaluation_challenges_fr: Vec<FsFr> = Vec::new();
+        let mut ys_fr: Vec<FsFr> = Vec::new();
+
+        for i in 0..blobs.len() {
+            let polynomial = blob_to_polynomial(&blobs[i]);
+            let evaluation_challenge_fr = compute_challenge(&blobs[i], &commitments_g1[i]);
+            let y_fr =
+                evaluate_polynomial_in_evaluation_form(&polynomial, &evaluation_challenge_fr, ks);
+
+            evaluation_challenges_fr.push(evaluation_challenge_fr);
+            ys_fr.push(y_fr);
+        }
+
+        verify_kzg_proof_batch(
+            commitments_g1,
+            &evaluation_challenges_fr,
+            &ys_fr,
+            proofs_g1,
+            ks,
+        )
+    }
 }