--- a/mcl/kzg/Cargo.toml
+++ b/mcl/kzg/Cargo.toml
@@ -14,6 +14,7 @@ primitive-types = "0.12.0"
 rayon = { version = "1.5.1", optional = true }
 once_cell = "1.4.0"
 libc = "0.2.139"
+num_cpus = "1.15.0"
 
 [features]
 parallel = ["rayon"]