--- a/zkcrypto/Cargo.toml
+++ b/zkcrypto/Cargo.toml
@@ -14,6 +14,7 @@ rand = "0.8"
 rand_core = "0.6"
 subtle = "2.2.1"
 rayon = "1.5.1"
+num_cpus = "1.15.0"
 
 [dependencies.ff]
 version = "0.12"