Below are the local benchmark results for `blst-from-scratch`:

| Benchmark | Old | New | Speedup |
|--------|--------|--------|--------|
| verify_blob_kzg_proof_batch/1   | 2.8333 ms | 2.8358 ms | 0.99x |
| verify_blob_kzg_proof_batch/2   | 4.6755 ms | 3.1930 ms | 1.46x |
| verify_blob_kzg_proof_batch/4   | 8.5825 ms | 3.2869 ms | 2.61x |
| verify_blob_kzg_proof_batch/8   | 15.849 ms | 6.2911 ms | 2.51x |
| verify_blob_kzg_proof_batch/16 | 29.943 ms | 12.389 ms | 2.41x |
| verify_blob_kzg_proof_batch/32 | 55.501 ms | 24.513 ms | 2.26x |
| verify_blob_kzg_proof_batch/64 | 108.27 ms | 48.739 ms | 2.22x |

Other implementations should yield similar results.