@sauliusgrigaitis 2023-04-17
Great work @belijzajac !

We need to try one more trick. IIRC we talked about this already during the call IIRC. If the number of blobs is higher than physical cores, then blobs should be split into groups and each individual group should run the batched algorithm (the one that c-kzg-4844 runs). For example, if there is 64 blobs and 4 cores, then there will be 4 groups 16 blobs each. Then each group runs on its own core. This should give a significant performance boost, especially for groups with more blobs.

You can get the number of physical cores by using [this](https://docs.rs/num_cpus/latest/num_cpus/fn.get_physical.html). You can also try to benchmark what happens if you split by logical cores. It should work slower.