@belijzajac 2023-04-17
So, I patched the `verify_kzg_proof_batch` test locally by changing the blobs count to an uneven number, specifically 47, also added a log printing right above the `blob_group.par_iter()` to see if the blobs splitting actually works, which indeed does:

<details><summary>Details</summary>
<p>

![Screenshot_20230417_150008](https://user-images.githubusercontent.com/36075948/232494913-10c28a23-df27-4e8e-8a56-87c6811e729d.png)

</p>
</details> 

And here are the local results:

| Benchmark | Old parallel | New parallel | Change |
|--------|--------|--------|--------|
| verify_blob_kzg_proof_batch/1   | 2.8333 ms | 2.8358 ms | +0.87% |
| verify_blob_kzg_proof_batch/2   | 4.6755 ms | 3.1930 ms | +1.34% |
| verify_blob_kzg_proof_batch/4   | 8.5825 ms | 3.2869 ms | +3.02% |
| verify_blob_kzg_proof_batch/8   | 15.849 ms | 6.2911 ms | +2.25% |
| verify_blob_kzg_proof_batch/16 | 29.943 ms | 12.389 ms | +2.16% |
| verify_blob_kzg_proof_batch/32 | 55.501 ms | 24.513 ms | +1.45% |
| verify_blob_kzg_proof_batch/64 | 108.27 ms | 48.739 ms | +0.89% |

This could be just a noise within benchmarks; however, I also replaced `map` to `all` method calls from the iterator trait because I found this in the documentation:

```text
all() is short-circuiting; in other words, it will stop processing as soon as it finds
a false, given that no matter what else happens, the result will also be false.
```

which guarantees an early exit if a single blob in the blob group is guaranteed to have an invalid proof. If we constructed a 4 blob groups containing 16 proofs each (64 blobs / 4 cores = 16 blobs and proofs) that evaluate to either true or false, as such:

```text
blob_group_1 [true, false, ...]
blob_group_2 [false, ...]
blob_group_3 [false, ...]
blob_group_4 [false, ...]
```

that would cost a total of 4 blob verification function calls!

This behavior is not shown in benchmarks because all the proofs generated in them are correct.