@sauliusgrigaitis 2023-04-17
For every group you need to call `verify_kzg_proof_batch` and pass the entire group. Again, let's say we have 2 cores and 4 blobs. This means on the first core there will be a single call to `verify_kzg_proof_batch` with blobs 0 and 1 passed, on the second core there will be a single `verify_kzg_proof_batch` call with blobs 2 and 3 passed.

Thanks not how it's currently implemented.