@belijzajac 2023-04-17
Comparison between current changes and the first commit:

| Benchmark | Old parallel (first commit) | New parallel (current commit) | Change |
|--------|--------|--------|--------|
| verify_blob_kzg_proof_batch/1   | 2.8203 ms | 2.8484 ms | +0.99%     |
| verify_blob_kzg_proof_batch/2   | 3.1175 ms | 4.7563 ms | **+52.56%** |
| verify_blob_kzg_proof_batch/4   | 3.2755 ms | 8.7011 ms | **+165.64%** |
| verify_blob_kzg_proof_batch/8   | 6.3278 ms | 5.3449 ms | -15.53%    |
| verify_blob_kzg_proof_batch/16 | 12.421 ms | 9.6115 ms | -22.61%    |
| verify_blob_kzg_proof_batch/32 | 24.551 ms | 17.584 ms | -28.37%    |
| verify_blob_kzg_proof_batch/64 | 48.766 ms | 33.336 ms | -31.64%    |

@sauliusgrigaitis should I include the parallel logic from the first commit when `num_blobs < num_cores`, taking into consideration that there doesn't exist such a cpu with 64 physical cores?