@belijzajac 2023-04-18
Comparison between current changes and the first commit:

| Benchmark | Old parallel (first commit) | New parallel (current commit) | Change |
|--------|--------|--------|--------|
| verify_blob_kzg_proof_batch/1   | 2.8358 ms | 2.8278 ms | -0.28%     |
| verify_blob_kzg_proof_batch/2   | 3.1478 ms | 3.2810 ms | **+4.23%** |
| verify_blob_kzg_proof_batch/4   | 3.2358 ms | 3.3705 ms | **+4.16%** |
| verify_blob_kzg_proof_batch/8   | 6.3014 ms | 5.3351 ms | - 15.33%    |
| verify_blob_kzg_proof_batch/16 | 12.327 ms | 9.6011 ms | -22.11%    |
| verify_blob_kzg_proof_batch/32 | 24.448 ms | 17.560 ms | -28.17%    |
| verify_blob_kzg_proof_batch/64 | 48.769 ms | 33.250 ms | -31.82%    |