@sauliusgrigaitis 2023-04-18
Great work @belijzajac . Slight degrade for 2 and 4 blobs is probably just a random?