@belijzajac 2023-04-18
I hope so, this is what I'm seeing in the CI benchmarks:

<details><summary>Details</summary>
<p>

![Screenshot_20230418_154643](https://user-images.githubusercontent.com/36075948/232782431-53fb4623-2498-4936-9c2d-5f63e25d5798.png)

</p>
</details>